migrate statementCreate function to annotator box
